+++ Source/Core/Geometry/Line3D.cs
@@ -60,11 +60,11 @@ namespace CodeImp.DoomBuilder.Geometry
 		public Vector3D GetDelta() { return End - Start; }
 
 		// This calculates the angle
-		public float GetAngle() 
+		public double GetAngle() 
 		{
 			// Calculate and return the angle
 			Vector2D d = GetDelta();
-			return -(float)Math.Atan2(-d.y, d.x) + Angle2D.PIHALF;
+			return -Math.Atan2(-d.y, d.x) + Angle2D.PIHALF;
 		}
 	}
 }
